add shceduler

+++ base/user.py
@@ -1,11 +1,11 @@
 from .base import metadata
 
-from sqlalchemy import Table, Column, Integer, String, DateTime
+from sqlalchemy import Table, Column, BigInteger, String, DateTime
 
 t_users = Table(
     "users",
     metadata,
-    Column('u_id', Integer),  # Идентификатор юзера в телеге
+    Column('u_id', BigInteger),  # Идентификатор юзера в телеге
     Column('fio', String),  # ФИО юзера
     Column('org', String),  # Организация пользователя
     Column('role', String),  # профиль прав пользователя
